requirements.txt
# Core Framework
streamlit>=1.31.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.1.0

# LLM & Embeddings  
groq>=0.4.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Speech Processing
openai-whisper>=20230918
sounddevice>=0.4.6    # For audio recording
pyaudio>=0.2.11       # Alternative audio backend

# Geolocation
geopy>=2.4.0
folium>=0.15.0        # Interactive maps

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
torch>=2.0.0          # For Whisper

# PDF Generation
fpdf2>=2.7.4

# Development
pytest>=7.4.0         # Testing framework
black>=23.0.0         # Code formatting
